Adds the api, storage and UI (form) implemented in the coding challenge
A first draft of Task 1 functionality. The user should be able to
interact with a text input field, entering any value to be sent to
the server for hashing.
Once correctly submitted, the server responds with a process id,
which can then be used to request the status of the hashing process
along with the hash value itself.
In these changes the api was misunderstood to already return the
hash value (as, for example, a transaction id). This will be corrected
in the next commit.