Cleans up previous commit implemented under time pressure
Corrects the misunderstanding that the /process endpoint server
response returns a hash, asserts that this is a process ID.

Makes types within the app more explicit and consistent. Fixes
naming of variables for consistency.

Adds tests for the two hash feature container components (we will
assume that in an existing application, shared presentational
components are already tested...). Tests focus on describing the
user experience - the integration between the UI and app data.